Inheritance of contact class
