+++ src/seedu/addressbook/data/person/Address.java
@@ -6,13 +6,13 @@ import seedu.addressbook.data.exception.IllegalValueException;
  * Represents a Person's address in the address book.
  * Guarantees: immutable; is valid as declared in {@link #isValidAddress(String)}
  */
-public class Address {
+public class Address extends Contact {
 
     public static final String EXAMPLE = "123, some street";
     public static final String MESSAGE_ADDRESS_CONSTRAINTS = "Person addresses can be in any format";
     public static final String ADDRESS_VALIDATION_REGEX = ".+";
 
-    public final String value;
+    //public final String value;
     private boolean isPrivate;
 
     /**
@@ -21,12 +21,14 @@ public class Address {
      * @throws IllegalValueException if given address string is invalid.
      */
     public Address(String address, boolean isPrivate) throws IllegalValueException {
-        String trimmedAddress = address.trim();
+        //String trimmedAddress = address.trim();
+        super(address);
         this.isPrivate = isPrivate;
-        if (!isValidAddress(trimmedAddress)) {
+        //if (!isValidAddress(trimmedAddress)) {
+        if (!isValidAddress(value)) {
             throw new IllegalValueException(MESSAGE_ADDRESS_CONSTRAINTS);
         }
-        this.value = trimmedAddress;
+        //this.value = trimmedAddress;
     }
 
     /**
@@ -36,10 +38,10 @@ public class Address {
         return test.matches(ADDRESS_VALIDATION_REGEX);
     }
 
-    @Override
-    public String toString() {
-        return value;
-    }
+    //@Override
+    //public String toString() {
+       // return value;
+    //}
 
     @Override
     public boolean equals(Object other) {
@@ -48,10 +50,10 @@ public class Address {
                 && this.value.equals(((Address) other).value)); // state check
     }
 
-    @Override
-    public int hashCode() {
-        return value.hashCode();
-    }
+    //@Override
+    //public int hashCode() {
+      //  return value.hashCode();
+    //}
 
     public boolean isPrivate() {
         return isPrivate;
