+++ src/seedu/addressbook/data/person/Email.java
@@ -6,14 +6,14 @@ import seedu.addressbook.data.exception.IllegalValueException;
  * Represents a Person's email in the address book.
  * Guarantees: immutable; is valid as declared in {@link #isValidEmail(String)}
  */
-public class Email {
+public class Email extends Contact {
 
     public static final String EXAMPLE = "valid@e.mail";
     public static final String MESSAGE_EMAIL_CONSTRAINTS =
             "Person emails should be 2 alphanumeric/period strings separated by '@'";
     public static final String EMAIL_VALIDATION_REGEX = "[\\w\\.]+@[\\w\\.]+";
 
-    public final String value;
+    //public final String value;
     private boolean isPrivate;
 
     /**
@@ -22,12 +22,14 @@ public class Email {
      * @throws IllegalValueException if given email address string is invalid.
      */
     public Email(String email, boolean isPrivate) throws IllegalValueException {
+        super(email);
         this.isPrivate = isPrivate;
-        String trimmedEmail = email.trim();
-        if (!isValidEmail(trimmedEmail)) {
+        //String trimmedEmail = email.trim();
+        //if (!isValidEmail(trimmedEmail)) {
+        if (!isValidEmail(value)) {
             throw new IllegalValueException(MESSAGE_EMAIL_CONSTRAINTS);
         }
-        this.value = trimmedEmail;
+        //this.value = trimmedEmail;
     }
 
     /**
@@ -37,10 +39,10 @@ public class Email {
         return test.matches(EMAIL_VALIDATION_REGEX);
     }
 
-    @Override
-    public String toString() {
-        return value;
-    }
+    //@Override
+    //public String toString() {
+      //  return value;
+    //}
 
     @Override
     public boolean equals(Object other) {
@@ -49,10 +51,10 @@ public class Email {
                 && this.value.equals(((Email) other).value)); // state check
     }
 
-    @Override
-    public int hashCode() {
-        return value.hashCode();
-    }
+    //@Override
+    //public int hashCode() {
+      //  return value.hashCode();
+    //}
 
 
     public boolean isPrivate() {
