+++ src/seedu/addressbook/data/person/Phone.java
@@ -6,13 +6,13 @@ import seedu.addressbook.data.exception.IllegalValueException;
  * Represents a Person's phone number in the address book.
  * Guarantees: immutable; is valid as declared in {@link #isValidPhone(String)}
  */
-public class Phone {
+public class Phone extends Contact {
 
     public static final String EXAMPLE = "123456789";
     public static final String MESSAGE_PHONE_CONSTRAINTS = "Person phone numbers should only contain numbers";
     public static final String PHONE_VALIDATION_REGEX = "\\d+";
 
-    public final String value;
+    //public final String value;
     private boolean isPrivate;
 
     /**
@@ -21,12 +21,14 @@ public class Phone {
      * @throws IllegalValueException if given phone string is invalid.
      */
     public Phone(String phone, boolean isPrivate) throws IllegalValueException {
+       super(phone);
         this.isPrivate = isPrivate;
-        String trimmedPhone = phone.trim();
-        if (!isValidPhone(trimmedPhone)) {
+        //String trimmedPhone = phone.trim();
+        //if (!isValidPhone(trimmedPhone)) {
+        if (!isValidPhone(value)) {
             throw new IllegalValueException(MESSAGE_PHONE_CONSTRAINTS);
         }
-        this.value = trimmedPhone;
+        //this.value = trimmedPhone;
     }
 
     /**
@@ -36,10 +38,10 @@ public class Phone {
         return test.matches(PHONE_VALIDATION_REGEX);
     }
 
-    @Override
-    public String toString() {
-        return value;
-    }
+    //@Override
+    //public String toString() {
+      //  return value;
+    //}
 
     @Override
     public boolean equals(Object other) {
@@ -48,10 +50,10 @@ public class Phone {
                 && this.value.equals(((Phone) other).value)); // state check
     }
 
-    @Override
-    public int hashCode() {
-        return value.hashCode();
-    }
+    //@Override
+    //public int hashCode() {
+      //  return value.hashCode();
+    //}
 
     public boolean isPrivate() {
         return isPrivate;
